refactor: merge b20-state-populate into base-state-populate with V1/V2 trie support
- Remove bin/b20-state-populate entirely; all functionality covered by the
  generic base-state-populate with --contract and --mapping-slot flags
- Add trie_version.rs: detects V1 (LegacyKeyAdapter) vs V2 (PackedKeyAdapter)
  from StorageSettings metadata at run time, so the tool works on both old
  mainnet snapshots and newer V2 datadirs without any flag
- Wire V1/V2 dispatch through compute_and_write_storage_trie,
  update_account_trie, and recompute_account_trie via TrieTableAdapter generics
- Fix critical: clear stale AccountsTrie nodes in recompute_account_trie before
  recomputing; without this the trie walker trusts cached hash flags, skips
  subtrees, and returns the old root on re-runs (mirrors the existing fix
  already applied to compute_and_write_storage_trie)
- Fix verify.rs: check_trie_nodes now V1/V2 adapter-aware; enforce
  AccountsTrie root presence with eyre::ensure! instead of just logging
- Fix check_balance_samples: sample indices now deduplicated and clamped to
  valid range (no OOB on count=1)
- Extract write_account_trie_nodes helper to avoid duplication between
  update_account_trie and recompute_account_trie
- Add serde_json + reth-storage-api deps (needed by trie_version detection)
