--- bin/b20-state-populate/Cargo.toml
@@ -1,39 +1,0 @@
-[package]
-name = "base-b20-state-populate"
-description = "B20 precompile and EVM ERC-20 state population testing utility"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-publish = false
-
-[lib]
-name = "base_b20_state_populate"
-path = "src/lib.rs"
-
-[[bin]]
-name = "b20-state-populate"
-path = "src/main.rs"
-
-[dependencies]
-eyre.workspace = true
-k256.workspace = true
-rand.workspace = true
-rayon.workspace = true
-tracing.workspace = true
-reth-trie.workspace = true
-indicatif.workspace = true
-reth-db-api.workspace = true
-reth-trie-db.workspace = true
-alloy-primitives.workspace = true
-serde_json.workspace = true
-reth-storage-api.workspace = true
-tracing-subscriber.workspace = true
-reth-primitives-traits.workspace = true
-clap = { workspace = true, features = ["derive"] }
-reth-db = { workspace = true, features = ["mdbx"] }
-
-[lints]
-workspace = true
