+++ bin/base-state-populate/Cargo.toml
@@ -24,9 +24,11 @@ rayon.workspace = true
 tracing.workspace = true
 reth-trie.workspace = true
 indicatif.workspace = true
+serde_json.workspace = true
 reth-db-api.workspace = true
 reth-trie-db.workspace = true
 alloy-primitives.workspace = true
+reth-storage-api.workspace = true
 tracing-subscriber.workspace = true
 reth-primitives-traits.workspace = true
 clap = { workspace = true, features = ["derive"] }
